+++ lib/turnip_formatter/template/index.erb
@@ -1,0 +1,65 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="UTF-8">
+    <title>turnip formatter report</title>
+    <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
+    <script src="http://ajax.googleapis.com/ajax/libs/jqueryui/1.10.2/jquery-ui.min.js"></script>
+    <script src="http://cdnjs.cloudflare.com/ajax/libs/jquery.tablesorter/2.9.1/jquery.tablesorter.min.js"></script>
+    <link rel="stylesheet" href="http://ajax.googleapis.com/ajax/libs/jqueryui/1.10.2/themes/smoothness/jquery-ui.css">
+
+    <style><%= TurnipFormatter::Template.css_render %></style>
+    <script><%= TurnipFormatter::Template.js_render %></script>
+  </head>
+  <body>
+    <div id="report">
+      <h1>Turnip Report</h1>
+      <section class="checkbox">
+        <label for="passed_check">Passed</label><input type="checkbox" id="passed_check">
+        <label for="failed_check">Failed</label><input type="checkbox" checked id="failed_check">
+        <label for="pending_check">Pending</label><input type="checkbox" id="pending_check">
+      </section>
+
+      <section class="result">
+        <p>
+          <span id="total_count"><%= h(scenarios.size) %></span> Scenario
+          (<span id="failed_count"><%= h(failed_count) %></span> failed
+          <span id="pending_count"><%= h(pending_count) %></span> pending).
+        </p>
+        <p>Finished in <span id="total_time"><%= h(total_time) %></span></p>
+      </section>
+    </div>
+
+    <div id="main" role="main">
+      <ul>
+        <li><a href="#steps-statistics">Steps</a></li>
+        <li><a href="#speed-statistics">Speed Statistics</a></li>
+        <li><a href="#feature-statistics">Feature Statistics</a></li>
+        <li><a href="#tag-statistics">Tag Statistics</a></li>
+      </ul>
+      <div id="steps-statistics">
+        <label><input type="checkbox" id="scenario_display_check" checked>step folding</label>
+        <% scenarios.each do |scenario| %>
+          <%= TurnipFormatter::Printer::Scenario.print_out(scenario) %>
+        <% end %>
+      </div>
+      <div id="feature-statistics">
+        <em>The results for the feature:</em>
+        <%= TurnipFormatter::Printer::TabFeatureStatistics.print_out(scenarios) %>
+      </div>
+      <div id="tag-statistics">
+        <em>The results for the tab:</em>
+        <%= TurnipFormatter::Printer::TabTagStatistics.print_out(scenarios) %>
+      </div>
+      <div id="speed-statistics">
+        <em>Ranking of running time of each <strong>successfully</strong> scenario:</em>
+        <%= TurnipFormatter::Printer::TabSpeedStatistics.print_out(passed_scenarios) %>
+      </div>
+    </div>
+
+    <footer>
+      <p>Generated by <a href="https://rubygems.org/gems/turnip_formatter">turnip_formatter</a> <%= TurnipFormatter::VERSION %></p>
+      <p>Powered by <a href="http://jquery.com/">jQuery</a> 1.9.1, <a href="http://jqueryui.com/">jQuery UI</a> 1.10.2 and <a href="http://mottie.github.io/tablesorter/">tablesorter</a> 2.9.1</p>
+    </footer>
+  </body>
+</html>
