+++ lib/turnip_formatter/template/runtime_exception.erb
@@ -1,0 +1,16 @@
+<section class="exception">
+  <h1>TurnipFormatter RuntimeError</h1>
+  <dl>
+    <%= runtime_exception %>
+
+    <dt>Example Full Description</dt>
+    <dd><%= h(example.metadata[:full_description]) %></dd>
+
+    <%= example_exception %>
+
+    <% if example.pending %>
+      <dt>Example Pending description</dt>
+      <dd><%= h(example.metadata[:description]) %></dd>
+    <% end %>
+  </dl>
+</section>
