+++ lib/turnip_formatter/template/scenario.erb
@@ -7,18 +7,19 @@
       Scenario: <%= h(scenario.name) %>
     </span>
     <span class="feature_name">
-      (Feature: <%= h(scenario.feature_info) %>)
-      at <%= h(scenario.run_time) %> sec
+      (Feature: <%= h(scenario.feature_info) %>) at <%= h(scenario.run_time) %> sec
     </span>
   </header>
-  <%= print_scenario_tags(scenario) %>
+
+  <ul class="tags">
+    <% scenario.tags.each do |tag| %>
+    <li>@<%= h(tag) %></li>
+    <% end %>
+  </ul>
+
   <ul class="steps">
     <% scenario.steps.each do |step| %>
-      <li <%= step_attr(step) %>><span><%= h(step.name) %></span>
-        <div class="args">
-          <%= step_args(step) %>
-        </div>
-      </li>
+      <%= TurnipFormatter::Printer::Step.print_out(step) %>
     <% end %>
   </ul>
 </section>
