--- lib/turnip_formatter/template/scenario_tags.erb
@@ -1,5 +1,0 @@
-<ul class="tags">
-  <% tags.each do |tag| %>
-  <li>@<%= h(tag) %></li>
-  <% end %>
-</ul>
