+++ lib/turnip_formatter/template/step.erb
@@ -1,0 +1,4 @@
+<li class="step <%= step.status %>">
+  <span><%= h(step.name) %></span>
+  <div class="args"><%= step_docs %></div>
+</li>
