+++ lib/turnip_formatter/template/tab_feature_statistics.erb
@@ -1,0 +1,24 @@
+<table>
+  <thead>
+    <tr>
+      <th>Feature</th>
+      <th>Scearios</th>
+      <th>passed</th>
+      <th>failed</th>
+      <th>pending</th>
+      <th>status</th>
+    </tr>
+  </thead>
+  <tbody>
+    <% analysis_results.each do |info| %>
+      <tr>
+        <td><%= info.name %></td>
+        <td><%= info.scenarios %></td>
+        <td><%= info.passed %></td>
+        <td><%= info.failed %></td>
+        <td><%= info.pending %></td>
+        <td class="<%= info.status %>"><%= info.status %></td>
+      </tr>
+    <% end %>
+  </tbody>
+</table>
