+++ lib/turnip_formatter/template/tab_speed_statistics.erb
@@ -1,0 +1,18 @@
+<table>
+  <thead>
+    <tr>
+      <th>Feature</th>
+      <th>Scenario</th>
+      <th>Duration</th>
+    </tr>
+  </thead>
+  <tbody>
+    <% analysis_results.each do |info| %>
+    <tr>
+      <td><span><%= info.feature_name %></span></td>
+      <td><a href="#<%= info.id %>"><%= info.name %></a></td>
+      <td><span><%= info.run_time %></span> sec</td>
+    </tr>
+    <% end %>
+  </tbody>
+</table>
