+++ spec/rspec/core/formatters/turnip_formatter_spec.rb
@@ -10,29 +10,17 @@ module RSpec::Core::Formatters
     let(:output) { StringIO.new }
     let(:formatter) { TurnipFormatter.new(output) }
 
-    describe '#start' do
-      it 'should be output header section' do
-        formatter.start(0)
-        expect(output.string).to match '<!DOCTYPE html>'
-        expect(output.string).to match '<div id="main" role="main">'
-      end
-    end
-
     describe '#example_passed' do
       before do
         scenario.example('passed', metadata) { expect(true).to be_true }
         feature.run(formatter)
       end
 
-      it 'should be output passed scenario section' do
-        string = output.string
-        expect(string).to match 'class="scenario passed"'
-      end
-
       it 'should get passed scenario count' do
         expect(formatter.passed_scenarios).to have(1).elements
         expect(formatter.failed_scenarios).to have(0).elements
         expect(formatter.pending_scenarios).to have(0).elements
+        expect(formatter.scenarios).to have(1).elements
       end
     end
 
@@ -49,14 +37,11 @@ module RSpec::Core::Formatters
         feature.run(formatter)
       end
 
-      it 'should be output failed scenario section' do
-        expect(output.string).to match 'class="scenario failed"'
-      end
-
       it 'should get failed scenario count' do
         expect(formatter.passed_scenarios).to have(0).elements
         expect(formatter.failed_scenarios).to have(1).elements
         expect(formatter.pending_scenarios).to have(0).elements
+        expect(formatter.scenarios).to have(1).elements
       end
     end
 
@@ -68,27 +53,25 @@ module RSpec::Core::Formatters
         feature.run(formatter)
       end
 
-      it 'should be output pending scenario section' do
-        expect(output.string).to match 'class="scenario pending"'
-      end
-
       it 'should get pending scenario count' do
         expect(formatter.passed_scenarios).to have(0).elements
         expect(formatter.failed_scenarios).to have(0).elements
         expect(formatter.pending_scenarios).to have(1).elements
+        expect(formatter.scenarios).to have(1).elements
       end
     end
 
     describe '#dump_summary' do
-      it 'should be output summary section' do
+      before do
         formatter.dump_summary(0.0, 3, 2, 1)
-        actual = output.string
+      end
 
-        expect(actual).to match %r{getElementById\("total_count"\).innerHTML = "3";}
-        expect(actual).to match %r{getElementById\("failed_count"\).innerHTML = "2";}
-        expect(actual).to match %r{getElementById\("pending_count"\).innerHTML = "1";}
-        expect(actual).to match %r{getElementById\("total_time"\).innerHTML = "0.0";}
-        expect(actual).to match '</html>'
+      subject { output.string }
+
+      it 'should be output summary section' do
+        should include '(<span id="failed_count">2</span> failed'
+        should include '<span id="pending_count">1</span> pending)'
+        should include '<span id="total_time">0.0</span>'
       end
     end
   end
